src/knowledge-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
psycopg[binary]>=3.2.0
psycopg-pool>=3.2.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
pyyaml>=6.0
structlog>=24.0.0
